Update lib to send vast error track
The RTB team needs to send VAST error tracks when there is a VAST
error. The vast client lib didn’t support that, so they had to add that
manually in the compiled file. It was a quick solution for a problem.
However the library itself is not the one in mbe-client, but a fork we
have. So any changes in mbe-client should be also implemented in the
fork. Because of that we had to put those changes in the original
coffee file and generate a new compiled file. Thus both mbe-client and
vast-client have the same code.