Update vast-client library with the latest change from original repository
Fyber has a fork for the vast-client library. The last time the library
was updated was some time ago and many bugs have been fixed since then.
Beyond that last year there was a request from product to add support
for multiple companion click tracking. So we had to do that in our fork
and create a [PR](dailymotion#109) to the original repository. The PR
was already merged and it is in master now. So it is recommendable that
we keep our fork up to date with the original repository to avoid
potential issues in the future.